+++ src/app/core/material/material.module.ts
@@ -4,8 +4,10 @@ import { MatCardModule } from '@angular/material/card';
 import { MatIconModule } from '@angular/material/icon';
 import { MatBadgeModule } from '@angular/material/badge';
 import { MatButtonModule } from '@angular/material/button';
+import { MatSelectModule } from '@angular/material/select';
 import { MatToolbarModule } from '@angular/material/toolbar';
 import { MatDividerModule } from '@angular/material/divider';
+import { MatCheckboxModule } from '@angular/material/checkbox';
 import { MatBottomSheetModule } from '@angular/material';
 
 const materialsModules = [
@@ -14,8 +16,10 @@ const materialsModules = [
   MatIconModule,
   MatBadgeModule,
   MatButtonModule,
+  MatSelectModule,
   MatToolbarModule,
   MatDividerModule,
+  MatCheckboxModule,
   MatBottomSheetModule
 ];